requirements.txt
# Predictive Maintenance Pipeline — Pinned Dependencies
# =====================================================
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
lightgbm==4.5.0
shap==0.46.0
pydantic==2.10.3
pydantic-settings==2.7.0
pyyaml==6.0.2
mlflow==2.19.0
joblib==1.4.2
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
prometheus-client==0.21.1
pytest==8.3.4
pytest-cov==6.0.0
ruff==0.8.4
mypy==1.13.0
pyarrow==18.1.0
